Extract renderOpenDropDownList as a function component from header
