+++ src/pages/hotnews.tsx
@@ -504,19 +504,19 @@ function HotNews() {
                 );
               })}
             </RestNewsDiv>
-            {isOpen && (
-              <Modal
-                content={hotNewsState[order].articleContent}
-                title={hotNewsState[order].title}
-                author={hotNewsState[order].author}
-                time={hotNewsState[order].publishedAt * 1000}
-                newsArticleUid={hotNewsState[order].id}
-                category={hotNewsState[order].category}
-                country={hotNewsState[order].country}
-                onClose={() => setIsOpen(false)}
-              />
-            )}
           </HotNewsBlock>
+        )}
+        {isOpen && (
+          <Modal
+            content={hotNewsState[order].articleContent}
+            title={hotNewsState[order].title}
+            author={hotNewsState[order].author}
+            time={hotNewsState[order].publishedAt * 1000}
+            newsArticleUid={hotNewsState[order].id}
+            category={hotNewsState[order].category}
+            country={hotNewsState[order].country}
+            onClose={() => setIsOpen(false)}
+          />
         )}
       </Wrapper>
     </Container>
